pyproject.toml
[project]
name = "shadowrun-lore-rag"
version = "0.1.0"
description = "RAG system for querying Shadowrun RPG rulebooks using natural language"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "marker-pdf>=1.10.2",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "langchain-text-splitters>=0.3.0",
]
